:truck: move complementaryCertificationRepo to `certification/configuration/`
