+++ api/src/certification/complementary-certification/domain/usecases/index.js
@@ -2,17 +2,15 @@
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import * as complementaryCertificationForTargetProfileAttachmentRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-for-target-profile-attachment-repository.js';
 import * as organizationRepository from '../../../complementary-certification/infrastructure/repositories/organization-repository.js';
+import * as complementaryCertificationRepository from '../../../configuration/infrastructure/repositories/complementary-certification-repository.js';
 import { mailService } from '../../../shared/domain/services/mail-service.js';
 import * as targetProfileHistoryRepository from '../../../shared/infrastructure/repositories/target-profile-history-repository.js';
 import * as complementaryCertificationBadgesRepository from '../../infrastructure/repositories/complementary-certification-badge-repository.js';
-import * as complementaryCertificationRepository from '../../infrastructure/repositories/complementary-certification-repository.js';
-
 /**
  *
  * Using {@link https://jsdoc.app/tags-type "Closure Compiler's syntax"} to document injected dependencies
  *
  * @typedef {complementaryCertificationBadgesRepository} ComplementaryCertificationBadgesRepository
- * @typedef {complementaryCertificationRepository} ComplementaryCertificationRepository
  * @typedef {complementaryCertificationForTargetProfileAttachmentRepository} ComplementaryCertificationForTargetProfileAttachmentRepository
  * @typedef {targetProfileHistoryRepository} TargetProfileHistoryRepository
  * @typedef {organizationRepository} OrganizationRepository
@@ -20,10 +18,10 @@ import * as complementaryCertificationRepository from '../../infrastructure/repo
  **/
 const dependencies = {
   complementaryCertificationBadgesRepository,
-  complementaryCertificationRepository,
   complementaryCertificationForTargetProfileAttachmentRepository,
   targetProfileHistoryRepository,
   organizationRepository,
+  complementaryCertificationRepository,
   mailService,
 };
