--- api/src/certification/shared/infrastructure/repositories/complementary-certification-repository.js
@@ -1,3 +1,0 @@
-import * as complementaryCertificationRepository from '../../../complementary-certification/infrastructure/repositories/complementary-certification-repository.js';
-
-export { complementaryCertificationRepository };
